Now times the function using high_res_clock instead of system_clock

+++ src/EngineBase.cpp
@@ -19,7 +19,7 @@ void EngineBase::addDrawAble(const std::shared_ptr<DrawAble>& drawAble)
     DrawAbleObjects::addDrawAble(drawAble);
 }
 
-void EngineBase::registerUpdateFunction(const std::function<void(float deltaTime)>& updateFunction)
+void EngineBase::registerUpdateFunction(const std::function<void(double deltaTime)>& updateFunction)
 {
     UpdateController::registerUpdateFunction(updateFunction);
 }
